Create the trading thread
Use an atomic on top of the buffer of `1`. Buffer of 1 means there can be another message in the queue while we're busy processing something else. Buffer of 0 not allowed. So the atomic is a safeguard to ensure we never have anything in the queue.

I want this, because I only want to execute trades that were allowed to start right away (no waiting). This is HFT: The prices will move.